Refactor components to fix max-lines-per-function lint rule
