+++ src/pages/ScheduledActions/index.tsx
@@ -1,7 +1,6 @@
 import { Button } from "@/components/Button";
 import { Card } from "@/components/Card";
-import { ConfirmDialog } from "@/components/ConfirmDialog";
-import { Pause, Pencil, Play, Plus, Trash } from "@/components/Icons";
+import { Plus } from "@/components/Icons";
 import {
 	useDeleteScheduledAction,
 	useInfiniteScheduledActionsList,
@@ -11,6 +10,8 @@ import React from "react";
 import { useNavigate } from "react-router-dom";
 import type { ScheduledAction } from "split-expense-shared-types";
 import styled, { useTheme } from "styled-components";
+import { useIntersectionObserver, useConfirmDialog } from "./hooks";
+import { ActionCard } from "./ActionCard";
 
 const HeaderTitle = styled.h3`
   margin: 0;
@@ -51,57 +52,6 @@ const ActionsGrid = styled.div`
   gap: 12px;
 `;
 
-const CardRow = styled.div`
-  display: flex;
-  justify-content: space-between;
-  gap: 12px;
-  align-items: center;
-`;
-
-const LeftSection = styled.div`
-  display: flex;
-  align-items: center;
-  gap: 10px;
-  cursor: pointer;
-`;
-
-const StatusDot = styled.span<{ $active: boolean }>`
-  width: 10px;
-  height: 10px;
-  border-radius: 50%;
-  background-color: ${(p) => (p.$active ? p.theme.colors.success : p.theme.colors.danger)};
-  box-shadow: 0 0 0 2px rgba(0, 0, 0, 0.03);
-  flex: 0 0 auto;
-`;
-
-const TitleText = styled.div`
-  font-weight: 600;
-  font-size: 16px;
-`;
-
-const Subtext = styled.div`
-  font-size: 12px;
-  color: ${({ theme }) => theme.colors.secondary};
-`;
-
-const Separator = styled.span`
-  margin: 0 6px;
-  color: ${({ theme }) => theme.colors.secondary};
-`;
-
-const RightActions = styled.div`
-  display: inline-flex;
-  align-items: center;
-  gap: 12px;
-`;
-
-const IconButton = styled.span`
-  cursor: pointer;
-  display: inline-flex;
-  align-items: center;
-  gap: 12px;
-`;
-
 const ScheduledActionsPage: React.FC = () => {
 	const navigate = useNavigate();
 	const {
@@ -119,45 +69,21 @@ const ScheduledActionsPage: React.FC = () => {
 	const actions: ScheduledAction[] =
 		data?.pages.flatMap((p) => p.scheduledActions) ?? [];
 
-	const sentinelRef = React.useRef<HTMLDivElement | null>(null);
-
-	React.useEffect(() => {
-		if (!sentinelRef.current) return;
-		const el = sentinelRef.current;
-		const observer = new IntersectionObserver(
-			(entries) => {
-				const entry = entries[0];
-				if (entry.isIntersecting && hasNextPage && !isFetchingNextPage) {
-					fetchNextPage();
-				}
-			},
-			{ root: null, rootMargin: "0px", threshold: 1.0 },
-		);
-		observer.observe(el);
-		return () => observer.unobserve(el);
-	}, [fetchNextPage, hasNextPage, isFetchingNextPage]);
-
-	const [confirmOpen, setConfirmOpen] = React.useState(false);
-	const [pendingDeleteId, setPendingDeleteId] = React.useState<string | null>(
-		null,
+	// Use custom hooks
+	const sentinelRef = useIntersectionObserver(
+		fetchNextPage,
+		hasNextPage,
+		isFetchingNextPage,
 	);
+	const { confirmOpen, pendingDeleteId, requestDelete, closeConfirm } =
+		useConfirmDialog();
 
-	const requestDelete = (id: string) => {
-		setPendingDeleteId(id);
-		setConfirmOpen(true);
-	};
-
+	// Create handlers
 	const confirmDelete = () => {
 		if (pendingDeleteId) {
 			deleteAction.mutate({ id: pendingDeleteId });
 		}
-		setConfirmOpen(false);
-		setPendingDeleteId(null);
-	};
-
-	const cancelDelete = () => {
-		setConfirmOpen(false);
-		setPendingDeleteId(null);
+		closeConfirm();
 	};
 
 	const theme = useTheme();
@@ -192,115 +118,18 @@ const ScheduledActionsPage: React.FC = () => {
 			{!isLoading && !isError && actions.length > 0 && (
 				<ActionsGrid>
 					{actions.map((sa) => (
-						<Card
+						<ActionCard
 							key={sa.id}
-							className="settings-card"
-							data-test-id={`sa-item-${sa.id}`}
-							style={
-								busyId === sa.id
-									? { opacity: 0.6, pointerEvents: "none" }
-									: undefined
-							}
-						>
-							<CardRow>
-								<LeftSection
-									onClick={() =>
-										navigate(`/scheduled-actions/${sa.id}`, { state: sa })
-									}
-								>
-									<StatusDot
-										$active={sa.isActive}
-										aria-label={sa.isActive ? "Active" : "Inactive"}
-										title={sa.isActive ? "Active" : "Inactive"}
-									/>
-									<div>
-										<TitleText>{sa.actionData.description}</TitleText>
-										<Subtext>
-											<span
-												style={{
-													textTransform: "uppercase",
-													letterSpacing: 0.3 as unknown as number,
-												}}
-											>
-												{sa.frequency}
-											</span>
-											<Separator>•</Separator>
-											<span>
-												{sa.actionType === "add_expense"
-													? "Add Expense"
-													: "Add to Budget"}
-											</span>
-											<Separator>•</Separator>
-											<span>Next: {sa.nextExecutionDate}</span>
-										</Subtext>
-									</div>
-								</LeftSection>
-								<RightActions>
-									<IconButton
-										onClick={async () => {
-											if (busyId) return;
-											setBusyId(sa.id);
-											try {
-												await updateAction.mutateAsync({
-													id: sa.id,
-													isActive: !sa.isActive,
-												});
-											} finally {
-												setBusyId(null);
-											}
-										}}
-										data-test-id={`sa-toggle-${sa.id}`}
-										aria-label={
-											sa.isActive ? "Deactivate action" : "Activate action"
-										}
-										title={
-											sa.isActive ? "Deactivate action" : "Activate action"
-										}
-									>
-										{busyId === sa.id ? (
-											<span aria-busy="true" style={{ opacity: 0.6 }}>
-												•••
-											</span>
-										) : sa.isActive ? (
-											<Pause size={18} color={theme.colors.danger} />
-										) : (
-											<Play size={18} color={theme.colors.success} />
-										)}
-									</IconButton>
-									<IconButton
-										onClick={() =>
-											navigate(`/scheduled-actions/${sa.id}/edit`, {
-												state: sa,
-											})
-										}
-										data-test-id={`sa-edit-${sa.id}`}
-										aria-label="Edit"
-										title="Edit"
-									>
-										<Pencil size={18} color={theme.colors.primary} outline />
-									</IconButton>
-									<IconButton
-										onClick={() => requestDelete(sa.id)}
-										data-test-id={`sa-delete-${sa.id}`}
-										aria-label="Delete"
-										title="Delete"
-									>
-										<Trash size={18} color={theme.colors.danger} />
-									</IconButton>
-								</RightActions>
-							</CardRow>
-							{confirmOpen && (
-								<ConfirmDialog
-									open={confirmOpen}
-									title="Delete action?"
-									message="This will permanently delete the scheduled action. This cannot be undone."
-									confirmText="Delete"
-									cancelText="Cancel"
-									onConfirm={confirmDelete}
-									onCancel={cancelDelete}
-								/>
-							)}
-						</Card>
+							sa={sa}
+							busyId={busyId}
+							setBusyId={setBusyId}
+							updateAction={updateAction}
+							requestDelete={requestDelete}
+							confirmOpen={confirmOpen}
+							pendingDeleteId={pendingDeleteId}
+							confirmDelete={confirmDelete}
+							closeConfirm={closeConfirm}
+						/>
 					))}
 					<div ref={sentinelRef} data-test-id="sa-infinite-sentinel" />
 					{isFetchingNextPage && (
